Do it in the workspace service
